disk_manager: fix disk codex.

+++ services/storage_daemon/disk/src/disk_manager.cpp
@@ -52,8 +52,8 @@ void DiskManager::HandleDiskEvent(NetlinkData *data)
         return;
     }
 
-    unsigned int major = std::stoi(data->GetParam("MAJOR"));
-    unsigned int minor = std::stoi(data->GetParam("MINOR"));
+    unsigned int major = (unsigned int) std::stoi(data->GetParam("MAJOR"));
+    unsigned int minor = (unsigned int) std::stoi(data->GetParam("MINOR"));
     dev_t device = makedev(major, minor);
 
     switch (data->GetAction()) {
@@ -88,8 +88,8 @@ std::shared_ptr<DiskInfo> DiskManager::MatchConfig(NetlinkData *data)
 {
     std::string sysPath = data->GetSyspath();
     std::string devPath = data->GetDevpath();
-    unsigned int major = std::stoi(data->GetParam("MAJOR"));
-    unsigned int minor = std::stoi(data->GetParam("MINOR"));
+    unsigned int major = (unsigned int) std::stoi(data->GetParam("MAJOR"));
+    unsigned int minor = (unsigned int) std::stoi(data->GetParam("MINOR"));
     dev_t device = makedev(major, minor);
 
     for (auto config : diskConfig_) {
